+++ tests/mocks.h
@@ -106,6 +106,7 @@ public:
     has_prog_kfunc_ = std::make_optional<bool>(has_features);
     prog_iter_task_ = std::make_optional<bool>(has_features);
     prog_iter_task_file_ = std::make_optional<bool>(has_features);
+    prog_iter_task_vma_ = std::make_optional<bool>(has_features);
     has_loop_ = std::make_optional<bool>(has_features);
     has_probe_read_kernel_ = std::make_optional<bool>(has_features);
     has_features_ = has_features;