+++ tests/runtime/engine/runner.py
@@ -131,6 +131,7 @@ class Runner(object):
         bpffeature["signal"] = output.find("send_signal: yes") != -1
         bpffeature["iter:task"] = output.find("iter:task: yes") != -1
         bpffeature["iter:task_file"] = output.find("iter:task_file: yes") != -1
+        bpffeature["iter:task_vma"] = output.find("iter:task_vma: yes") != -1
         bpffeature["libpath_resolv"] = output.find("bcc library path resolution: yes") != -1
         bpffeature["dwarf"] = output.find("libdw (DWARF support): yes") != -1
         bpffeature["kprobe_multi"] = output.find("kprobe_multi: yes") != -1