+++ src/attached_probe.h
@@ -43,7 +43,7 @@ private:
   std::string eventname() const;
   static std::string sanitise(const std::string &str);
   void resolve_offset_kprobe(bool safe_mode);
-  void resolve_offset_uprobe(bool safe_mode);
+  bool resolve_offset_uprobe(bool safe_mode);
   void load_prog(BPFfeature &feature);
   void attach_multi_kprobe(void);
   void attach_kprobe(bool safe_mode);