+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to
   - [#2521](https://github.com/iovisor/bpftrace/pull/2521)
 - Respect BPFTRACE_STRLEN environment variable for all strings
   - [#2545](https://github.com/iovisor/bpftrace/pull/2545)
+- Treat str() builtin's len parameter as int64
+  - [#2546](https://github.com/iovisor/bpftrace/pull/2546)
 #### Docs
 #### Tools