+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to
   - [#2557](https://github.com/iovisor/bpftrace/pull/2557)
 - BTF support for tracepoints defined in modules
   - [#2479](https://github.com/iovisor/bpftrace/pull/2479)
+- Add trailer to truncated strings
+  - [#2559](https://github.com/iovisor/bpftrace/pull/2559)
 #### Changed
 - Improve attaching to uprobes with size 0
   - [#2562](https://github.com/iovisor/bpftrace/pull/2562)